+++ R/airsis_createMetaDataframe.R
@@ -130,7 +130,8 @@ airsis_createMetaDataframe <- function(tbl,
   }
   
   # Add elevation, siteName and countyName
-  meta <- addGoogleMetadata(meta, existingMeta=existingMeta)
+  meta <- addGoogleElevation(meta, existingMeta=existingMeta)
+  meta <- addGoogleAddress(meta, existingMeta=existingMeta)
   
   # Convert some columns to character even if they have all NA
   characterColumns <- c('siteName','agencyName','countyName','msaName','monitorType',